+++ packages/contracts/contracts/test/invariant/actors/DepositManagerHandler.sol
@@ -33,7 +33,7 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
     string constant CONTRACT_VERSION = "v1";
 
     uint256 constant ETH_SUPPLY = 120_500_000 ether;
-    uint256 constant GAS_COMPENSATION = 100_000 * 50 gwei;
+    uint256 constant GAS_COMPENSATION_PER_DEPOSIT = 120_000 gwei;
 
     uint256 constant SCREENER_PRIVKEY = 1;
     address SCREENER_ADDRESS = vm.addr(SCREENER_PRIVKEY);
@@ -46,14 +46,11 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
     SimpleERC721Token public erc721;
     SimpleERC1155Token public erc1155;
 
-    EncodedAsset public encodedErc20;
-
     bytes32 public lastCall;
 
     // ______INTERNAL______
     mapping(bytes32 => uint256) internal _calls;
     mapping(string => uint256) internal _reverts;
-    uint256[] internal _depositSizes;
     AddressSet internal _actors;
     address internal _currentActor;
     uint256 internal _actorNum = 0;
@@ -78,12 +75,6 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
         erc20 = _erc20;
         erc721 = _erc721;
         erc1155 = _erc1155;
-
-        encodedErc20 = AssetUtils.encodeAsset(
-            AssetType.ERC20,
-            address(_erc20),
-            ERC20_ID
-        );
     }
 
     modifier createActor() {
@@ -146,75 +137,128 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
     }
 
     function instantiateDepositETH(
-        uint256 amount
+        uint256 seed
     ) public createActor trackCall("instantiateDepositETH") {
-        // Bound deposit amount
-        amount = bound(amount, 0, ETH_SUPPLY);
-        _depositSizes.push(amount);
+        uint256 numDeposits = bound(seed, 1, 10);
 
-        // Deal gas compensation
-        vm.deal(_currentActor, amount + GAS_COMPENSATION);
-
-        vm.startPrank(_currentActor);
-
-        // Start recording logs and make call
-        vm.recordLogs();
+        // Bound deposit amounts, save deposit reqs
+        uint256[] memory depositAmounts = new uint256[](numDeposits);
+        DepositRequest[] memory deposits = new DepositRequest[](numDeposits);
         StealthAddress memory depositAddr = NocturneUtils
             .defaultStealthAddress();
-        depositManager.instantiateETHDeposit{value: amount + GAS_COMPENSATION}(
-            amount,
-            depositAddr
+        for (uint256 i = 0; i < numDeposits; i++) {
+            // Get random amount
+            uint256 newSeed;
+            unchecked {
+                newSeed = seed + i;
+            }
+            uint256 amount = bound(
+                uint256(keccak256(abi.encodePacked(newSeed))),
+                0,
+                ETH_SUPPLY
+            );
+            depositAmounts[i] = amount;
+
+            // Record deposit req
+            deposits[i] = NocturneUtils.formatDepositRequest(
+                _currentActor,
+                address(erc20),
+                depositAmounts[i],
+                NocturneUtils.ERC20_ID,
+                depositAddr,
+                depositManager._nonce() + i,
+                GAS_COMPENSATION_PER_DEPOSIT
+            );
+        }
+
+        // Deal enough eth for deposits + comp
+        uint256 totalDepositAmount = _sum(depositAmounts);
+        vm.deal(
+            _currentActor,
+            totalDepositAmount + (GAS_COMPENSATION_PER_DEPOSIT * numDeposits)
         );
 
-        // Recover deposit request
-        Vm.Log[] memory entries = vm.getRecordedLogs();
-        Vm.Log memory entry = entries[entries.length - 1];
-        DepositRequest memory req = EventParsing
-            .decodeDepositRequestFromDepositEvent(entry);
-
-        vm.stopPrank();
+        vm.prank(_currentActor);
+        depositManager.instantiateETHMultiDeposit{
+            value: totalDepositAmount +
+                (GAS_COMPENSATION_PER_DEPOSIT * numDeposits)
+        }(depositAmounts, depositAddr);
 
         // Update sets and sum
-        _depositSet.push(req);
-        _instantiateDepositSumSetETH.addToActorSum(_currentActor, amount);
+        for (uint256 i = 0; i < numDeposits; i++) {
+            _depositSet.push(deposits[i]);
+            _instantiateDepositSumSetETH.addToActorSum(
+                _currentActor,
+                depositAmounts[i]
+            );
+        }
     }
 
     function instantiateDepositErc20(
-        uint256 amount
+        uint256 numDepositsSeed,
+        uint256 seed
     ) public createActor trackCall("instantiateDepositErc20") {
-        // Bound deposit amount
-        amount = bound(amount, 0, type(uint256).max - erc20.totalSupply());
-        erc20.reserveTokens(_currentActor, amount);
-        _depositSizes.push(amount);
+        uint256 numDeposits = bound(numDepositsSeed, 1, 10);
+
+        (, uint32 globalCapWholeTokens, , , uint8 precision) = depositManager
+            ._erc20Caps(address(erc20));
+        uint256 globalCap = uint256(globalCapWholeTokens) * 10 ** precision;
+
+        // Bound deposit amounts, reserve tokens, save deposit reqs
+        uint256[] memory depositAmounts = new uint256[](numDeposits);
+        DepositRequest[] memory deposits = new DepositRequest[](numDeposits);
+        StealthAddress memory depositAddr = NocturneUtils
+            .defaultStealthAddress();
+        for (uint256 i = 0; i < numDeposits; i++) {
+            // Get random amount
+            uint256 newSeed;
+            unchecked {
+                newSeed = seed + i;
+            }
+            uint256 amount = bound(
+                uint256(keccak256(abi.encodePacked(newSeed))),
+                0,
+                globalCap
+            );
+            depositAmounts[i] = amount;
+
+            // Record deposit req
+            deposits[i] = NocturneUtils.formatDepositRequest(
+                _currentActor,
+                address(erc20),
+                depositAmounts[i],
+                NocturneUtils.ERC20_ID,
+                depositAddr,
+                depositManager._nonce() + i,
+                GAS_COMPENSATION_PER_DEPOSIT
+            );
+        }
+
+        // Reserve tokens
+        uint256 totalDepositAmount = _sum(depositAmounts);
+        erc20.reserveTokens(_currentActor, totalDepositAmount);
 
         // Deal gas compensation
-        vm.deal(_currentActor, GAS_COMPENSATION);
+        vm.deal(_currentActor, GAS_COMPENSATION_PER_DEPOSIT * numDeposits);
 
         // Approve token
         vm.startPrank(_currentActor);
-        erc20.approve(address(depositManager), amount);
+        erc20.approve(address(depositManager), totalDepositAmount);
 
-        // Start recording logs and make call
-        vm.recordLogs();
-        StealthAddress memory depositAddr = NocturneUtils
-            .defaultStealthAddress();
-        depositManager.instantiateDeposit{value: GAS_COMPENSATION}(
-            encodedErc20,
-            amount,
-            depositAddr
-        );
-
-        // Recover deposit request
-        Vm.Log[] memory entries = vm.getRecordedLogs();
-        Vm.Log memory entry = entries[entries.length - 1];
-        DepositRequest memory req = EventParsing
-            .decodeDepositRequestFromDepositEvent(entry);
+        depositManager.instantiateErc20MultiDeposit{
+            value: GAS_COMPENSATION_PER_DEPOSIT * numDeposits
+        }(address(erc20), depositAmounts, depositAddr);
 
         vm.stopPrank();
 
-        // Update sets and sum
-        _depositSet.push(req);
-        _instantiateDepositSumSetErc20.addToActorSum(_currentActor, amount);
+        // Update deposit set and sum
+        for (uint256 i = 0; i < numDeposits; i++) {
+            _depositSet.push(deposits[i]);
+            _instantiateDepositSumSetErc20.addToActorSum(
+                _currentActor,
+                depositAmounts[i]
+            );
+        }
     }
 
     function retrieveDepositErc20(
@@ -233,10 +277,14 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
         // Retrieve deposit
         vm.prank(randDepositRequest.spender);
         try depositManager.retrieveDeposit(randDepositRequest) {
-            (uint256 encodedWethAddr, ) = depositManager._wethEncoded();
+            EncodedAsset memory encodedWeth = AssetUtils.encodeAsset(
+                AssetType.ERC20,
+                address(depositManager._weth()),
+                ERC20_ID
+            );
             if (
                 randDepositRequest.encodedAsset.encodedAssetAddr ==
-                encodedWethAddr
+                encodedWeth.encodedAssetAddr
             ) {
                 _retrieveDepositSumSetETH.addToActorSum(
                     randDepositRequest.spender,
@@ -276,11 +324,20 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
         );
 
         // Complete deposit
-        try depositManager.completeDeposit(randDepositRequest, signature) {
-            (uint256 encodedWethAddr, ) = depositManager._wethEncoded();
+        uint256 warpTimestamp;
+        unchecked {
+            warpTimestamp = block.timestamp + seed;
+        }
+        vm.warp(warpTimestamp);
+        try depositManager.completeErc20Deposit(randDepositRequest, signature) {
+            EncodedAsset memory encodedWeth = AssetUtils.encodeAsset(
+                AssetType.ERC20,
+                address(depositManager._weth()),
+                ERC20_ID
+            );
             if (
                 randDepositRequest.encodedAsset.encodedAssetAddr ==
-                encodedWethAddr
+                encodedWeth.encodedAssetAddr
             ) {
                 _completeDepositSumSetETH.addToActorSum(
                     randDepositRequest.spender,
@@ -363,5 +420,14 @@ contract DepositManagerHandler is CommonBase, StdCheats, StdUtils {
         address actor
     ) public view returns (uint256) {
         return _completeDepositSumSetErc20.getSumForActor(actor);
+    }
+
+    // ______PURE______
+    function _sum(uint256[] memory arr) internal pure returns (uint256) {
+        uint256 sum = 0;
+        for (uint256 i = 0; i < arr.length; i++) {
+            sum += arr[i];
+        }
+        return sum;
     }
 }